Add files via upload
This repository contains a Python implementation of the K-Means Clustering algorithm for classifying Iris species. The Iris dataset is a popular dataset in machine learning and consists of three different species of Iris flowers: Setosa, Versicolor, and Virginica.

K-Means Clustering is an unsupervised machine learning algorithm used for clustering data points into groups or clusters based on their similarities. In this project, K-Means is applied to cluster Iris flowers into different groups based on their characteristics.